Density, pressure fix
Particles densities and pressure calculations now work as intended, also made some changes to the camera so it's more visible.
The target density was high and the particles tried to collapse on the 0,0,0, it was a weird thing.
I have to limit also the target density to a slider from 0.1 to 1f, because it crashes the simulation once it pases a value of ~1.5